move optimize helpers to -util rather than -operation
