+++ packages/client/hmi-client/src/components/workflow/ops/optimize-ciemss/optimize-ciemss-operation.ts
@@ -1,14 +1,7 @@
 import { Operation, WorkflowOperationTypes, BaseState } from '@/types/workflow';
-import {
-	Intervention,
-	InterventionSemanticType,
-	InterventionPolicy,
-	StaticIntervention,
-	DynamicIntervention
-} from '@/types/Types';
-import { CiemssMethodOptions, getRunResult, getSimulation } from '@/services/models/simulation-service';
-import { getModelIdFromModelConfigurationId } from '@/services/model-configurations';
-import { createInterventionPolicy, blankIntervention } from '@/services/intervention-policy';
+import { StaticIntervention, DynamicIntervention } from '@/types/Types';
+import { CiemssMethodOptions } from '@/services/models/simulation-service';
+import { blankIntervention } from '@/services/intervention-policy';
 import optimizeModel from '@assets/svg/operator-images/optimize-model.svg';
 import { ChartSetting } from '@/types/common';
 
@@ -177,128 +170,3 @@ export const OptimizeCiemssOperation: Operation = {
 		return init;
 	}
 };
-
-// Get the simulation object that ran the optimize.
-// Get the fixed static interventions from this, as well as some details about the optimization interevntions
-// Get the optimization result file
-// Concat the optimization result file with the optimization interventions from simulation object
-export async function getOptimizedInterventions(optimizeRunId: string) {
-	// This is a snake case version of types/Types OptimizeInterventions as we are reading from simulation execution payload
-	interface OptimizeInterventionsSnakeCase {
-		intervention_type: string;
-		param_name: string;
-		param_value?: number;
-		start_time?: number;
-		time_objective_function: string;
-		parameter_objective_function: string;
-		relative_importance: number;
-		param_value_initial_guess?: number;
-		parameter_value_lower_bound: number;
-		parameter_value_upper_bound: number;
-		start_time_initial_guess?: number;
-		start_time_lower_bound: number;
-		start_time_upper_bound: number;
-	}
-
-	const allInterventions: Intervention[] = [];
-	// Get the interventionPolicyGroups from the simulation object.
-	// This will prevent any inconsistencies being passed via knobs or state when matching with result file.
-	const simulation = await getSimulation(optimizeRunId);
-
-	const simulationStaticInterventions: any[] = simulation?.executionPayload.fixed_static_parameter_interventions ?? [];
-	const optimizeInterventions: OptimizeInterventionsSnakeCase[] = simulation?.executionPayload?.optimize_interventions;
-
-	// Add any static (not optimized) interventions to allInterventions
-	simulationStaticInterventions.forEach((inter) => {
-		const newIntervetion: Intervention = {
-			dynamicInterventions: inter.dynamic_interventions,
-			name: inter.name,
-			staticInterventions: inter.static_interventions
-		};
-		allInterventions.push(newIntervetion);
-	});
-
-	const policyResults: number[] = await getRunResult(optimizeRunId, 'policy.json');
-
-	optimizeInterventions.forEach((optimizedIntervention) => {
-		const interventionType = optimizedIntervention.intervention_type;
-		const paramName: string = optimizedIntervention.param_name;
-		const paramValue: number | undefined = optimizedIntervention.param_value;
-		const startTime: number | undefined = optimizedIntervention.start_time;
-
-		if (interventionType === OptimizationInterventionObjective.startTime) {
-			const newTimestepAsList = policyResults.splice(0, 1);
-			// If we our intervention type is param value our policyResult will provide a timestep.
-			allInterventions.push({
-				name: `Optimized ${paramName}`,
-				staticInterventions: [
-					{
-						appliedTo: paramName,
-						type: InterventionSemanticType.Parameter,
-						timestep: newTimestepAsList[0],
-						value: paramValue as number
-					}
-				],
-				dynamicInterventions: []
-			});
-		} else if (interventionType === OptimizationInterventionObjective.paramValue) {
-			const valueAsList = policyResults.splice(0, 1);
-			// If we our intervention type is start time our policyResult will provide a parameter value.
-			allInterventions.push({
-				name: `Optimized ${paramName}`,
-				staticInterventions: [
-					{
-						timestep: startTime as number,
-						value: valueAsList[0],
-						appliedTo: paramName,
-						type: InterventionSemanticType.Parameter
-					}
-				],
-				dynamicInterventions: []
-			});
-		} else if (interventionType === OptimizationInterventionObjective.paramValueAndStartTime) {
-			// If our intervention type is start_time_param_value our policyResult will contain the timestep value, then the parameter value.
-			// https://github.com/ciemss/pyciemss/blob/main/pyciemss/integration_utils/intervention_builder.py#L66
-			const timeAndValueAsList = policyResults.splice(0, 2);
-			allInterventions.push({
-				name: `Optimized ${paramName}`,
-				staticInterventions: [
-					{
-						timestep: timeAndValueAsList[0],
-						value: timeAndValueAsList[1],
-						appliedTo: paramName,
-						type: InterventionSemanticType.Parameter
-					}
-				],
-				dynamicInterventions: []
-			});
-		} else {
-			// Should realistically not be hit unless we change the interface and do not update
-			console.error(`Unable to find the intevention type for optimization run: ${optimizeRunId}`);
-		}
-	});
-	return allInterventions;
-}
-
-/**
- *
- * 1) Get optimize interventions
- *
- *
- */
-export async function createInterventionPolicyFromOptimize(
-	modelConfigId: string,
-	optimizeRunId: string
-): Promise<InterventionPolicy | null> {
-	const modelId = await getModelIdFromModelConfigurationId(modelConfigId);
-	const optimizedInterventions = await getOptimizedInterventions(optimizeRunId);
-
-	const newIntervention: InterventionPolicy = {
-		name: `Optimize run: ${optimizeRunId}`,
-		modelId,
-		temporary: true,
-		interventions: optimizedInterventions
-	};
-	const newInterventionPolicy: InterventionPolicy | null = await createInterventionPolicy(newIntervention);
-	return newInterventionPolicy;
-}
